* 'master' of https://gitee.com/landun_tech/MultiEducation:
  when in babyContentScene, load newest babyPhoto
  optimize babyArchives judge condition
  1. break_point_from_endpoint_upload func implementation but not test 2. sove recurrence func bug in babyPhotoProfile
  optimize babyArchives judge conditions
  optimize babyArchives judge conditions

+++ MultiEducation/PBFiles/MebabyGrowth.pbobjc.h
@@ -124,10 +124,10 @@ typedef GPB_ENUM(GuStudentArchivesPb_FieldNumber) {
 @property(nonatomic, readwrite, copy, null_resettable) NSString *zodiac;
 
 /** 身高 */
-@property(nonatomic, readwrite) int32_t height;
+@property(nonatomic, readwrite) float height;
 
 /** 体重 */
-@property(nonatomic, readwrite) int32_t weight;
+@property(nonatomic, readwrite) float weight;
 
 /** 左眼视力 */
 @property(nonatomic, readwrite) float leftVision;
